Debug notification for users logging on and off.
Previously, a notification would be generated for each user who was logged in. Now, it only occurs once for everyone, regardless of number of users. This was because I had coupled the userCount handler (which updates the App state, which is passed as a prop to the navbar) with the notification. Since each user needs to have their state updated each time a user logs on/off, each user was then sending a message event to the server, so the server would receive as many messages as users.

By creating a second object and a second send to the client (but this time only to the last client in the Set, which I converted to an array), I was able to get things working the way I want. Thanks Vaz!